refactor(conversion): merge pipeline batch façade into core engine
The batch entry points (convert_documents, ConversionBundle, LaTeXFragment,
to_template_fragments) lived in a separate pipeline.py that only looped over
convert_document and wrapped its results — a thin layer over the single-document
engine in core.py. Fold them into core.py and delete pipeline.py so the
conversion engine lives in one module. Importers (top-level texsmith package,
ConversionService, TemplateSession, CLI presenter, tests) now import from
texsmith.core.conversion.core; the public texsmith.* surface is unchanged.

Combined with the prior context-threading cleanup, the per-job orchestration is
now service -> core (convert_documents -> convert_document -> render) instead of
service -> pipeline -> core -> render.

All 849 tests green, ruff clean.
