+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and the project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Extracted the babel / BCP-47 language tables and mappers out of the 1039-line `texsmith.core.templates.manifest` into a focused `texsmith.core.templates.languages` module (manifest down to ~930 lines); `runtime` now imports the alias table from there instead of reaching into `manifest`.
 - Moved the two heading-analysis `HTMLParser` subclasses out of `texsmith.core.documents` (637 → ~547 lines) into `texsmith.core.heading_analysis` (`HeadingLevelScanner`, `HeadingInspector`), and dropped the unused `Document._HEADING_TAGS` class variable.
 - Split the LaTeX-log Rich renderer (`LatexLogRenderer`) out of the 720-line `texsmith.adapters.latex.engines.latex.log` into a sibling `log_render` module (log down to ~500 lines), separating log *parsing* from its *presentation*; the streaming entry point imports the renderer lazily to avoid an import cycle.
+- Collapsed the conversion-orchestration layering: the batch façade (`convert_documents`, `ConversionBundle`, `LaTeXFragment`, `to_template_fragments`) moved from `texsmith.core.conversion.pipeline` into `texsmith.core.conversion.core` (which already owned the single-document engine), and `pipeline.py` was removed — one fewer module/hop for the same job. The top-level `texsmith.convert_documents` / `ConversionBundle` / `LaTeXFragment` public API is unchanged; only the deep `texsmith.core.conversion.pipeline` import path is gone (use `texsmith.core.conversion.core`). Internally, `_render_document` / `_build_runtime_common` now read the document, request flags and generation strategy off the `ConversionContext` instead of receiving them as a long list of unpacked keyword arguments.
 
 ### Removed
