+++ REFACTOR.md
@@ -61,7 +61,7 @@ Commits (sur `refactor/ir-migration`) : `7e3233d` (Lots 1–3 + glossary), `1b63
 ## Lot 4 — Refonte architecturale (risque élevé)
 ### Reste à faire — chantiers structurants (par ordre de valeur/risque)
 
-1. **Collapser les 4 couches d'orchestration** → `service.py` (façade) + `core.py` (moteur) ; absorber la boucle de `pipeline.py` + `ConversionBundle`/`LaTeXFragment` ; supprimer l'indirection `convert_document` mono-doc. *(~120 LOC, touche le cœur — exige run complet à chaque étape.)*
+1. [x] **Collapse orchestration** — fait en 2 temps : (A) `_render_document`/`_build_runtime_common` lisent `document`/`request`/`generation` depuis le `context` au lieu de 11 kwargs threadés ; (B) `pipeline.py` (API batch + `ConversionBundle`/`LaTeXFragment`/`convert_documents`/`to_template_fragments`) fusionné dans `core.py` et supprimé — moteur de conversion désormais en un seul module. Importeurs repointés vers `conversion.core` ; API publique `texsmith.*` inchangée. *(Note : `convert_document` mono-doc conservé — fusionner dans `_render_document` créerait une fonction de ~250 L, anti-lecture.)*
 2. **Sortir l'orchestration moteur LaTeX** de `ui/cli/commands/render.py:~1065-1216` vers un `core.build_pdf(...)` (logique métier hors CLI ; supprime les hooks monkeypatch `render.shutil/subprocess/run_engine_command`).
 3. **Relocaliser `render_typst_document`** (`typst_emit.py`) dans `core/conversion` ; promouvoir les 5 imports privés (`_load_inline_bibliography`, `_build_mustache_defaults`, `_replace_mustaches_in_html`, `_resolve_callout_style`, `_citation_label`) en API publique — supprime le pipeline Typst parallèle.
 4. **Fermer la fuite IR** : lower `exercise`/`snippet`/`rich_table` dans le reader (plus de `BeautifulSoup`/`HtmlReader` rappelés depuis les writers, `writer.py:~1173-1294`).
