+++ src/texsmith/core/conversion/core.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from collections.abc import Callable, Mapping
+from collections.abc import Callable, Iterable, Mapping, Sequence
 import copy
 import dataclasses
 from dataclasses import dataclass, field
@@ -25,7 +25,8 @@ from texsmith.core.templates import (
     wrap_template_document,
 )
 
-from ..diagnostics import DiagnosticEmitter
+from ..diagnostics import DiagnosticEmitter, NullEmitter
+from ._utils import build_unique_stem_map
 from .debug import (
     debug_enabled,
     ensure_emitter,
@@ -36,6 +37,7 @@ from .debug import (
 )
 from .execution import resolve_conversion_context
 from .models import ConversionRequest
+from .renderer import TemplateFragment
 from .templates import (
     SlotFragment,
     bind_template,
@@ -641,9 +643,146 @@ def render_with_fallback(
     return output, current_state
 
 
+@dataclass(slots=True)
+class LaTeXFragment:
+    """Represents a rendered LaTeX fragment."""
+
+    document: Document
+    latex: str
+    stem: str
+    output_path: Path | None = None
+    conversion: ConversionResult | None = None
+
+    def write_to(self, target: Path) -> None:
+        """Persist the fragment to disk so later template steps can reference it."""
+        target.parent.mkdir(parents=True, exist_ok=True)
+        target.write_text(self.latex, encoding="utf-8")
+        self.output_path = target
+
+
+@dataclass(slots=True)
+class ConversionBundle:
+    """Collection returned by :func:`convert_documents`."""
+
+    fragments: list[LaTeXFragment]
+
+    def combined_output(self) -> str:
+        """Concatenate all fragments separated by blank lines for quick previews."""
+        return "\n\n".join(fragment.latex for fragment in self.fragments if fragment.latex)
+
+
+def convert_documents(
+    documents: Sequence[Document],
+    *,
+    output_dir: Path | None = None,
+    settings: ConversionRequest | None = None,
+    emitter: DiagnosticEmitter | None = None,
+    bibliography_files: Iterable[Path] | None = None,
+    template: str | None = None,
+    template_runtime: TemplateRuntime | None = None,
+    template_overrides: Mapping[str, Any] | None = None,
+    wrap_document: bool = True,
+    shared_state: DocumentState | None = None,
+    write_fragments: bool | None = None,
+) -> ConversionBundle:
+    """Convert one or more documents into LaTeX fragments while coordinating shared state."""
+    if not documents:
+        raise ValueError("At least one document is required for conversion.")
+
+    request = settings.copy() if settings is not None else ConversionRequest()
+    request.bibliography_files = list(bibliography_files or [])
+    request.template = template
+    unique_stems = build_unique_stem_map([doc.source_path for doc in documents])
+
+    shared_bibliography: BibliographyCollection | None = None
+    seen_bibliography_issues: set[tuple[str, str | None, str | None]] = set()
+
+    if bibliography_files:
+        shared_bibliography = BibliographyCollection()
+        shared_bibliography.load_files(bibliography_files)
+
+    fragments: list[LaTeXFragment] = []
+    should_write_fragments = write_fragments if write_fragments is not None else True
+    state = shared_state
+    active_emitter = emitter or NullEmitter()
+
+    for document in documents:
+        document = document.prepare_for_conversion()
+        target_dir = Path(output_dir) if output_dir is not None else Path("build")
+        slot_overrides = dict(document.slot_selectors)
+        result = convert_document(
+            document=document,
+            output_dir=target_dir,
+            request=request,
+            slot_overrides=slot_overrides or None,
+            emitter=active_emitter,
+            template_overrides=template_overrides,
+            state=None if wrap_document else state,
+            template_runtime=template_runtime,
+            wrap_document=wrap_document,
+            preloaded_bibliography=shared_bibliography,
+            seen_bibliography_issues=seen_bibliography_issues,
+        )
+
+        if not wrap_document:
+            state = result.document_state or state
+
+        stem = unique_stems[document.source_path]
+        fragment = LaTeXFragment(
+            document=document,
+            latex=result.latex_output,
+            stem=stem,
+            conversion=result,
+        )
+        if output_dir is not None and should_write_fragments:
+            target = target_dir / f"{stem}.tex"
+            fragment.write_to(target)
+        fragments.append(fragment)
+
+    return ConversionBundle(fragments=fragments)
+
+
+def to_template_fragments(bundle: ConversionBundle) -> list[TemplateFragment]:
+    """Convert bundle fragments into the template fragment contract used by the template engine."""
+    fragments: list[TemplateFragment] = []
+    for fragment in bundle.fragments:
+        conversion = fragment.conversion
+        if conversion is None:
+            raise ValueError("Template rendering requires conversion metadata for each fragment.")
+
+        document = fragment.document
+        slot_includes: set[str] = set()
+        if document is not None:
+            slot_includes = set(document.slot_includes)
+
+        fragments.append(
+            TemplateFragment(
+                stem=fragment.stem,
+                latex=conversion.latex_output,
+                default_slot=conversion.default_slot or "mainmatter",
+                slot_outputs=dict(conversion.slot_outputs),
+                slot_includes=slot_includes,
+                document_state=conversion.document_state,
+                bibliography_path=conversion.bibliography_path,
+                template_engine=conversion.template_engine,
+                requires_shell_escape=conversion.template_shell_escape,
+                template_overrides=dict(conversion.template_overrides),
+                output_path=fragment.output_path,
+                front_matter=document.front_matter,
+                source_path=document.source_path,
+                assets=dict(conversion.assets_map),
+            )
+        )
+    return fragments
+
+
 __all__ = [
+    "ConversionBundle",
     "ConversionResult",
+    "LaTeXFragment",
     "convert_document",
+    "convert_documents",
     "copy_document_state",
     "render_with_fallback",
+    "to_template_fragments",
 ]
